Implements pokemon Gym parsing
